+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [4.2.10] - 2022-10-28
+### Added
+- Prelimiary work on Research Projects [#2473](https://github.com/rokwire/illinois-app/issues/2473).
+### Changed
+- Text, layout and navigation updates for Research Questionnaire [#2484](https://github.com/rokwire/illinois-app/issues/2484).
+
 ## [4.2.9] - 2022-10-27
 ### Changed
 - Store research questionnaire answers in account profile [#2477](https://github.com/rokwire/illinois-app/issues/2477).